fix export type
