+++ packages/superset-ui-translation/src/index.ts
@@ -1,4 +1,2 @@
-import '../types/external.d';
-
 export { configure, t, tn } from './TranslatorSingleton';
 export { TranslatorConfig, LanguagePack } from './Translator';
