+++ packages/superset-ui-translation/test/languagePacks/en.ts
@@ -1,4 +1,4 @@
-import { LanguagePack } from 'jed';
+import { LanguagePack } from '../../src/index';
 
 const languagePack: LanguagePack = {
   domain: 'superset',
